Version 0.9.8.1

+++ lib/mp4v2/mp4file.cpp
@@ -2104,7 +2104,7 @@ u_int64_t MP4File::ConvertFromTrackTimestamp(
 	MP4Timestamp timeStamp,
 	u_int32_t timeScale)
 {
-	return MP4ConvertTime((u_int64_t)timeStamp, 
+	return MP4ConvertTime(timeStamp, 
 		GetTrackTimeScale(trackId), timeScale);
 }
